refactor(chat): split project conversation slices
Summary:
- move project conversation lifecycle, runtime orchestration, and
  stream projection flows into dedicated chat files
- extract workspace file path and content parsing into stable domain
  types and keep the edit helpers focused on mutation workflows
- split project conversation HTTP handlers out of `chat_api.go` and
  add focused regression tests for lifecycle, projection, and
  workspace file parsing

Rationale:
- reduce the blast radius of Project Conversation changes by keeping
  lifecycle, runtime, workspace, and streaming concerns on narrower
  internal boundaries
- keep handlers in parse/auth/delegate shape while preserving the
  existing external API and stream semantics
- lock in the refactor with automated coverage on the new slice
  boundaries instead of relying on manual verification

Tests:
- PATH=$HOME/.local/go1.26.1/bin:$PATH OPENASE_PGTEST_SHARED_ROOT=$HOME/.cache/openase/pgtest go test ./internal/chat ./internal/httpapi
